+++ src/Telegram.Bot/Types/ReactionType.cs
@@ -1,0 +1,62 @@
+using Telegram.Bot.Converters;
+
+namespace Telegram.Bot.Types;
+
+/// <summary>
+/// This object describes the type of a reaction. Currently, it can be one of
+/// <list type="bullet">
+/// <item><description>ReactionTypeEmoji</description></item>
+/// <item><description>ReactionTypeCustomEmoji</description></item>
+/// </list>
+/// </summary>
+[JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
+[JsonConverter(typeof(ReactionTypeConverter))]
+public abstract class ReactionType
+{
+    /// <summary>
+    /// Type of the reaction
+    /// </summary>
+    [JsonProperty]
+    public abstract string Type { get; }
+}
+
+/// <summary>
+/// The reaction is based on an emoji.
+/// </summary>
+[JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
+public class ReactionTypeEmoji : ReactionType
+{
+    /// <summary>
+    /// Type of the reaction, always "emoji"
+    /// </summary>
+    public override string Type => "emoji";
+
+    /// <summary>
+    /// Reaction emoji. Currently, it can be one of "👍", "👎", "❤", "🔥", "🥰", "👏", "😁",
+    /// "🤔", "🤯", "😱", "🤬", "😢", "🎉", "🤩", "🤮", "💩", "🙏", "👌", "🕊", "🤡", "🥱",
+    /// "🥴", "😍", "🐳", "❤‍🔥", "🌚", "🌭", "💯", "🤣", "⚡", "🍌", "🏆", "💔", "🤨",
+    /// "😐", "🍓", "🍾", "💋", "🖕", "😈", "😴", "😭", "🤓", "👻", "👨‍💻", "👀", "🎃",
+    /// "🙈", "😇", "😨", "🤝", "✍", "🤗", "🫡", "🎅", "🎄", "☃", "💅", "🤪", "🗿", "🆒",
+    /// "💘", "🙉", "🦄", "😘", "💊", "🙊", "😎", "👾", "🤷‍♂", "🤷", "🤷‍♀", "😡"
+    /// </summary>
+    [JsonProperty(Required = Required.Always)]
+    public string Emoji { get; } = default!;
+}
+
+/// <summary>
+/// The reaction is based on an emoji.
+/// </summary>
+[JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
+public class ReactionTypeCustomEmoji : ReactionType
+{
+    /// <summary>
+    /// Type of the reaction, always "custom_emoji"
+    /// </summary>
+    public override string Type => "custom_emoji";
+
+    /// <summary>
+    /// Custom emoji identifier
+    /// </summary>
+    [JsonProperty(Required = Required.Always)]
+    public string CustomEmojiId { get; } = default!;
+}