+++ test/Telegram.Bot.Tests.Unit/EnumConverter/UpdateTypeConverterTests.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Telegram.Bot.Types.Enums;
@@ -8,22 +10,25 @@ namespace Telegram.Bot.Tests.Unit.EnumConverter;
 
 public class UpdateTypeConverterTests
 {
+    [Fact]
+    public void Should_Verify_All_UpdateType_Members()
+    {
+        List<string> updateTypeMembers = Enum
+            .GetNames<UpdateType>()
+            .OrderBy(x => x)
+            .ToList();
+        List<string> updateTypeDataMembers = new UpdateTypeData()
+            .Select(x => Enum.GetName(typeof(UpdateType), x[0]))
+            .OrderBy(x => x)
+            .ToList()!;
+
+        Assert.Equal(updateTypeMembers.Count, updateTypeDataMembers.Count);
+        Assert.Equal(updateTypeMembers, updateTypeDataMembers);
+    }
+
+
     [Theory]
-    [InlineData(UpdateType.Unknown, "unknown")]
-    [InlineData(UpdateType.Message, "message")]
-    [InlineData(UpdateType.InlineQuery, "inline_query")]
-    [InlineData(UpdateType.ChosenInlineResult, "chosen_inline_result")]
-    [InlineData(UpdateType.CallbackQuery, "callback_query")]
-    [InlineData(UpdateType.EditedMessage, "edited_message")]
-    [InlineData(UpdateType.ChannelPost, "channel_post")]
-    [InlineData(UpdateType.EditedChannelPost, "edited_channel_post")]
-    [InlineData(UpdateType.ShippingQuery, "shipping_query")]
-    [InlineData(UpdateType.PreCheckoutQuery, "pre_checkout_query")]
-    [InlineData(UpdateType.Poll, "poll")]
-    [InlineData(UpdateType.PollAnswer, "poll_answer")]
-    [InlineData(UpdateType.MyChatMember, "my_chat_member")]
-    [InlineData(UpdateType.ChatMember, "chat_member")]
-    [InlineData(UpdateType.ChatJoinRequest, "chat_join_request")]
+    [ClassData(typeof(UpdateTypeData))]
     public void Should_Convert_UpdateType_To_String(UpdateType updateType, string value)
     {
         Update update = new(updateType);
@@ -38,7 +43,7 @@ public class UpdateTypeConverterTests
     }
 
     [Theory]
-    [InlineData(UpdateType.Unknown, "unknown")]
+    [ClassData(typeof(UpdateTypeData))]
     public void Should_Convert_String_To_UpdateType(UpdateType updateType, string value)
     {
         Update expectedResult = new(updateType);
@@ -77,4 +82,30 @@ public class UpdateTypeConverterTests
 
     [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
     record Update([property: JsonProperty(Required = Required.Always)] UpdateType Type);
+
+    private class UpdateTypeData : IEnumerable<object[]>
+    {
+        public IEnumerator<object[]> GetEnumerator()
+        {
+            yield return new object[] { UpdateType.Unknown, "unknown" };
+            yield return new object[] { UpdateType.Message, "message" };
+            yield return new object[] { UpdateType.InlineQuery, "inline_query" };
+            yield return new object[] { UpdateType.ChosenInlineResult, "chosen_inline_result" };
+            yield return new object[] { UpdateType.CallbackQuery, "callback_query" };
+            yield return new object[] { UpdateType.EditedMessage, "edited_message" };
+            yield return new object[] { UpdateType.ChannelPost, "channel_post" };
+            yield return new object[] { UpdateType.EditedChannelPost, "edited_channel_post" };
+            yield return new object[] { UpdateType.ShippingQuery, "shipping_query" };
+            yield return new object[] { UpdateType.PreCheckoutQuery, "pre_checkout_query" };
+            yield return new object[] { UpdateType.Poll, "poll" };
+            yield return new object[] { UpdateType.PollAnswer, "poll_answer" };
+            yield return new object[] { UpdateType.MyChatMember, "my_chat_member" };
+            yield return new object[] { UpdateType.ChatMember, "chat_member" };
+            yield return new object[] { UpdateType.ChatJoinRequest, "chat_join_request" };
+            yield return new object[] { UpdateType.MessageReaction, "message_reaction" };
+            yield return new object[] { UpdateType.MessageReactionCount, "message_reaction_count" };
+        }
+
+        System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => GetEnumerator();
+    }
 }