+++ CHANGELOG.md
@@ -25,6 +25,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 > [Bot API 7.0](https://core.telegram.org/bots/api#december-29-2023) (December 29, 2023)
 
+### Added
+
+- The classes `ReactionType`, `ReactionTypeEmoji` and `ReactionTypeCustomEmoji` representing different types of reaction.
+- Updates about a reaction change on a message with non-anonymous reactions, represented by the class `MessageReactionUpdated`
+and the field `MessageReaction` in the class `Update`. The bot must explicitly allow the update to receive it.
+- Updates about reaction changes on a message with anonymous reactions, represented by the class `MessageReactionCountUpdated`
+and the field `MessageReactionCount` in the class `Update`. The bot must explicitly allow the update to receive it.
+- New enum values `MessageReaction`, `MessageReactionCount` for `UpdateType`.
+- Type `ReactionCount`.
+- Request type `SetMessageReactionRequest` that allows bots to react to messages.
+- New method `ITelegramBotClient.SetMessageReactionAsync` that allows bots to react to messages.
+- The field `AvailableReactions` to the class `Chat`.
+
 ## [v20.0.0] - Unreleased
 
 > [Bot API 6.9](https://core.telegram.org/bots/api#september-22-2023) (September 22, 2023)